Created the payment status update use case

+++ src/main/java/com/github/barbosaluc/testefadesp/dto/PagamentoRequestDTO.java
@@ -2,8 +2,11 @@ package com.github.barbosaluc.testefadesp.dto;
 
 import com.github.barbosaluc.testefadesp.domain.enums.MetodoPagamento;
 
+import com.github.barbosaluc.testefadesp.domain.enums.StatusPagamento;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
+
+import java.math.BigDecimal;
 
 public record PagamentoRequestDTO(
 
@@ -16,6 +19,6 @@ public record PagamentoRequestDTO(
     Long numeroCartao,
 
     @NotNull(message = "O valor do pagamento é obrigatório")
-    Double valorPagamento
+    BigDecimal valorPagamento
 
 ) {}
